% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONEST.R
\docType{data}
\name{empirical}
\alias{empirical}
\title{Precomputed dataset}
\usage{
data('empirical')
}
\description{
A precomputed dataset only used by the ONEST_vignettes function to create the vignettes.
}
\details{
\itemize{
  \item columns : different permutations
  \item rows : agreement percentage
}
}
\keyword{datasets}
