% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONEST.R
\docType{data}
\name{NCCN_22c3_t}
\alias{NCCN_22c3_t}
\title{Example dataset5}
\usage{
data('NCCN_22c3_t')
}
\description{
NCCN 22c3 tumor data from Rimm et al. (2017) used to illustrate the ONEST method.
}
\details{
\itemize{
  \item columns : different raters
  \item rows : different cases
}
}
\references{
[1] Rimm DL, Han G, Taube JM, Yi ES, Bridge JA, Flieder DB, et al. A Prospective, Multi-institutional, Pathologist-Based Assessment of 4 Immunohistochemistry Assays for PD-L1 Expression in Non-Small Cell Lung Cancer. JAMA Oncol. 2017;3(8):1051-8, DOI: 10.1001/jamaoncol.2017.0013, PMID: 28278348.
}
\keyword{datasets}
