% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONEST_inflation_test.R
\name{ONEST_inflation_test}
\alias{ONEST_inflation_test}
\title{The ONEST_inflation_test}
\usage{
ONEST_inflation_test(Y)
}
\arguments{
\item{Y}{pathology data as a matrix (missing values are allowed)}
}
\value{
p_value: chi-square p-value, a small p-value indicates significant evidence that the observers’ agreement will converge to a non-zero proportion.
}
\description{
This function uses the binary (0/1) pathology data to test if the agreement proportion will converge to 0 with an increasing number of observers.
}
\examples{
data("sp263_bin")
ONEST_inflation_test(sp263_bin)
}
