% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONEST.R
\docType{data}
\name{sp263_bin}
\alias{sp263_bin}
\title{Example dataset2}
\usage{
data('sp263_bin')
}
\description{
A pathology reads dataset of triple negative breast cancer in Reisenbichler et al. (2020). A value of 0 means negative and a value of 1 means positive.
}
\details{
\itemize{
  \item columns : different raters
  \item rows : different cases
}
}
\references{
[1] Reisenbichler ES, Han G, Bellizzi A, Bossuyt V, Brock J, Cole K, et al. Prospective multi-institutional evaluation of pathologist assessment of PD-L1 assays for patient selection in triple negative breast cancer. Mod Pathol. 2020, DOI: 10.1038/s41379-020-0544-x; PMID: 32300181.
}
\keyword{datasets}
